Actualización del menú de la Agenda

+++ agenda/src/com/generation/ejercicio/informacion/Contacto.java
@@ -3,9 +3,9 @@ package com.generation.ejercicio.informacion;
 public class Contacto {
     protected String name;
     protected String apellido;
-    protected int telefono;
+    protected long telefono;
 
-    public Contacto(String apellido, String name, int telefono) {
+    public Contacto(String apellido, String name, long telefono) {
         this.apellido = apellido;
         this.name = name;
         this.telefono = telefono;
@@ -27,7 +27,7 @@ public class Contacto {
         this.name = name;
     }
 
-    public int getTelefono() {
+    public long getTelefono() {
         return telefono;
     }
